+++ Audit.NET.sln
@@ -148,6 +148,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Audit.AzureStorageBlobs.Uni
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Audit.DynamoDB.UnitTest", "test\Audit.DynamoDB.UnitTest\Audit.DynamoDB.UnitTest.csproj", "{C0186428-77E9-4317-AE5E-08141D8CADAF}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Audit.Elasticsearch.UnitTest", "test\Audit.Elasticsearch.UnitTest\Audit.Elasticsearch.UnitTest.csproj", "{04D5CCB9-E095-40D3-9986-C1E41D8F638F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -636,6 +638,14 @@ Global
 		{C0186428-77E9-4317-AE5E-08141D8CADAF}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C0186428-77E9-4317-AE5E-08141D8CADAF}.Release|x64.ActiveCfg = Release|Any CPU
 		{C0186428-77E9-4317-AE5E-08141D8CADAF}.Release|x64.Build.0 = Release|Any CPU
+		{04D5CCB9-E095-40D3-9986-C1E41D8F638F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{04D5CCB9-E095-40D3-9986-C1E41D8F638F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{04D5CCB9-E095-40D3-9986-C1E41D8F638F}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{04D5CCB9-E095-40D3-9986-C1E41D8F638F}.Debug|x64.Build.0 = Debug|Any CPU
+		{04D5CCB9-E095-40D3-9986-C1E41D8F638F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{04D5CCB9-E095-40D3-9986-C1E41D8F638F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{04D5CCB9-E095-40D3-9986-C1E41D8F638F}.Release|x64.ActiveCfg = Release|Any CPU
+		{04D5CCB9-E095-40D3-9986-C1E41D8F638F}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -703,6 +713,7 @@ Global
 		{C4EEAE98-1596-4177-B821-308DEA359A78} = {A54B4BB6-3439-432B-AFD9-FE62D6528D42}
 		{2167C9A6-432A-4B44-AFD1-8B71BCE940B9} = {A54B4BB6-3439-432B-AFD9-FE62D6528D42}
 		{C0186428-77E9-4317-AE5E-08141D8CADAF} = {A54B4BB6-3439-432B-AFD9-FE62D6528D42}
+		{04D5CCB9-E095-40D3-9986-C1E41D8F638F} = {A54B4BB6-3439-432B-AFD9-FE62D6528D42}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {B1989475-43D6-4AA5-9717-6DBF734B0C6E}
