+++ test/Run-Tests.ps1
@@ -43,6 +43,8 @@ StartDotnetUnitTests 'Audit.IntegrationTest' 'MySql' '--filter=TestCategory=MySq
 StartDotnetUnitTests 'Audit.IntegrationTest' 'PostgreSQL' '--filter=TestCategory=PostgreSQL';
 StartDotnetUnitTests 'Audit.AzureCosmos.UnitTest' 'AzureCosmos';
 StartDotnetUnitTests 'Audit.AzureStorageBlobs.UnitTest' 'AzureStorageBlobs';
+StartDotnetUnitTests 'Audit.DynamoDB.UnitTest' 'DynamoDB';
+StartDotnetUnitTests 'Audit.Elasticsearch.UnitTest' 'Elasticsearch';
 # StartDotnetUnitTests 'Audit.IntegrationTest' 'AzureStorage' '--filter=TestCategory=AzureBlob|TestCategory=AzureStorageBlobs|TestCategory=AzureTables';
 # StartDotnetUnitTests 'Audit.AzureStorageTables.UnitTest' 'AzureTables';
 StartDotnetUnitTests 'Audit.IntegrationTest' 'Elasticsearch' '--filter=TestCategory=Elasticsearch';
